Move common dedup utils and remove unused code (#42)
* Refactor common utils and remove unused code


* More cleanup


* More updates/shuffling


* Move gpu_dedup scripts into subfolder


* Remove gpu_deduplication subfolder


* Add readme to fuzzy dedup scripts section


* Fix typo and relative links


* Remove legacy script entrypoints


* Remove legacy scripts and add init file


* Update GpuDeduplication.rst


---------
