+++ nemo_curator/scripts/fuzzy_deduplication/jaccard_compute.py
@@ -15,9 +15,9 @@
 import os
 import time
 
-from nemo_curator.gpu_deduplication.utils import parse_nc_args
 from nemo_curator.modules.fuzzy_dedup import JaccardSimilarity
 from nemo_curator.utils.distributed_utils import get_client, get_num_workers
+from nemo_curator.utils.script_utils import parse_gpu_dedup_args
 
 
 def main(args):
@@ -57,7 +57,7 @@ def main(args):
 def attach_args(parser=None):
     description = """Computes jaccard similarity"""
     if not parser:
-        parser = parse_nc_args(description=description)
+        parser = parse_gpu_dedup_args(description=description)
 
     parser.add_argument(
         "--shuffled-docs-path",
