+++ setup.py
@@ -89,18 +89,13 @@ setup(
             "prepare_task_data=nemo_curator.scripts.prepare_task_data:console_script",
             "find_matching_ngrams=nemo_curator.scripts.find_matching_ngrams:console_script",
             "remove_matching_ngrams=nemo_curator.scripts.remove_matching_ngrams:console_script",
-            "gpu_compute_minhashes=nemo_curator.scripts.compute_minhashes:console_script",
-            "minhash_buckets=nemo_curator.scripts.minhash_lsh:console_script",
-            "jaccard_map_buckets=nemo_curator.scripts.map_buckets:console_script",
-            "jaccard_shuffle=nemo_curator.scripts.jaccard_shuffle:console_script",
-            "jaccard_compute=nemo_curator.scripts.jaccard_compute:console_script",
-            "gpu_connected_component=nemo_curator.scripts.connected_components:console_script",
-            "write_deduped_result_with_text=nemo_curator.gpu_deduplication.write_deduped_result_with_text:console_script",
-            "verify_all_pairs_jaccard=nemo_curator.gpu_deduplication.verify_all_pairs_jaccard:console_script",
+            "gpu_compute_minhashes=nemo_curator.scripts.fuzzy_deduplication.compute_minhashes:console_script",
+            "minhash_buckets=nemo_curator.scripts.fuzzy_deduplication.minhash_lsh:console_script",
+            "jaccard_map_buckets=nemo_curator.scripts.fuzzy_deduplication.map_buckets:console_script",
+            "jaccard_shuffle=nemo_curator.scripts.fuzzy_deduplication.jaccard_shuffle:console_script",
+            "jaccard_compute=nemo_curator.scripts.fuzzy_deduplication.jaccard_compute:console_script",
+            "gpu_connected_component=nemo_curator.scripts.fuzzy_deduplication.connected_components:console_script",
             "gpu_exact_dups=nemo_curator.scripts.find_exact_duplicates:console_script",
-            "prepare_fuzzy_ids=nemo_curator.gpu_deduplication.prepare_fuzzy_ids:console_script",
-            "create_list_of_duplicate_ids=nemo_curator.gpu_deduplication.create_list_of_duplicate_ids:console_script",
-            "remove_duplicates=nemo_curator.gpu_deduplication.remove_duplicates:console_script",
             "deidentify=nemo_curator.scripts.find_pii_and_deidentify:console_script",
             "generate_statistics=nemo_curator.distributed_data_classification.generate_statistics:console_script",
             "domain_classifier_inference=nemo_curator.distributed_data_classification.domain_classifier_inference:console_script",
